+++ src/symbol/collision_index.ts
@@ -150,7 +150,8 @@ class CollisionIndex {
             lineVertexArray,
             labelPlaneMatrix,
             projectionCache,
-            getElevation);
+            getElevation,
+            false);
 
         let collisionDetected = false;
         let inGrid = false;